+++ src/ExternalSearch.Providers.DnB.Provider/DnBProviderProviderComponent.cs
@@ -3,50 +3,48 @@ using Castle.MicroKernel.Registration;
 using CluedIn.Core;
 using CluedIn.Core.Providers;
 using CluedIn.Core.Server;
-using CluedIn.ExternalSearch.Providers.DnB;
 using ComponentHost;
 using Constants = CluedIn.ExternalSearch.Providers.DnB.DnBConstants;
 
-namespace CluedIn.Provider.DnB
+namespace CluedIn.Provider.ExternalSearch.DnB;
+
+[Component(Constants.ComponentName, "Providers", ComponentType.Service, ServerComponents.ProviderWebApi, Components.Server, Components.DataStores, Isolation = ComponentIsolation.NotIsolated)]
+public sealed class DnBProviderProviderComponent : ServiceApplicationComponent<IServer>
 {
-    [Component(Constants.ComponentName, "Providers", ComponentType.Service, ServerComponents.ProviderWebApi, Components.Server, Components.DataStores, Isolation = ComponentIsolation.NotIsolated)]
-    public sealed class DnBProviderProviderComponent : ServiceApplicationComponent<IServer>
+    /**********************************************************************************************************
+     * CONSTRUCTOR
+     **********************************************************************************************************/
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="DnBProviderProviderComponent" /> class.
+    /// </summary>
+    /// <param name="componentInfo">The component information.</param>
+    public DnBProviderProviderComponent(ComponentInfo componentInfo) : base(componentInfo)
     {
-        /**********************************************************************************************************
-         * CONSTRUCTOR
-         **********************************************************************************************************/
+        // Dev. Note: Potential for compiler warning here ... CA2214: Do not call overridable methods in constructors
+        //   this class has been sealed to prevent the CA2214 waring being raised by the compiler
+        Container.Register(Component.For<DnBProviderProviderComponent>().Instance(this));
+    }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="GoogleMapsProviderProviderComponent" /> class.
-        /// </summary>
-        /// <param name="componentInfo">The component information.</param>
-        public DnBProviderProviderComponent(ComponentInfo componentInfo) : base(componentInfo)
-        {
-            // Dev. Note: Potential for compiler warning here ... CA2214: Do not call overridable methods in constructors
-            //   this class has been sealed to prevent the CA2214 waring being raised by the compiler
-            Container.Register(Component.For<DnBProviderProviderComponent>().Instance(this));
-        }
+    /**********************************************************************************************************
+     * METHODS
+     **********************************************************************************************************/
 
-        /**********************************************************************************************************
-         * METHODS
-         **********************************************************************************************************/
+    /// <summary>Starts this instance.</summary>
+    public override void Start()
+    {
+        var asm = Assembly.GetAssembly(typeof(DnBProviderProviderComponent));
+        Container.Register(Types.FromAssembly(asm).BasedOn<IProvider>().WithServiceFromInterface().If(t => !t.IsAbstract).LifestyleSingleton());
 
-        /// <summary>Starts this instance.</summary>
-        public override void Start()
-        {
-            var asm = Assembly.GetAssembly(typeof(DnBProviderProviderComponent));
-            Container.Register(Types.FromAssembly(asm).BasedOn<IProvider>().WithServiceFromInterface().If(t => !t.IsAbstract).LifestyleSingleton());
+        State = ServiceState.Started;
+    }
 
-            State = ServiceState.Started;
-        }
+    /// <summary>Stops this instance.</summary>
+    public override void Stop()
+    {
+        if (State == ServiceState.Stopped)
+            return;
 
-        /// <summary>Stops this instance.</summary>
-        public override void Stop()
-        {
-            if (State == ServiceState.Stopped)
-                return;
-
-            State = ServiceState.Stopped;
-        }
+        State = ServiceState.Stopped;
     }
 }
