+++ src/ExternalSearch.Providers.DnB.Provider/DnBSearchProviderProvider.cs
@@ -1,4 +1,8 @@
-﻿using CluedIn.Core;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using CluedIn.Core;
 using CluedIn.Core.Crawling;
 using CluedIn.Core.Data.Relational;
 using CluedIn.Core.ExternalSearch;
@@ -8,132 +12,127 @@ using CluedIn.ExternalSearch;
 using CluedIn.ExternalSearch.Providers.DnB;
 using CluedIn.ExternalSearch.Providers.DnB.Vocabularies;
 using CluedIn.Providers.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Constants = CluedIn.ExternalSearch.Providers.DnB.DnBConstants;
 
-namespace CluedIn.Provider.DnB
+namespace CluedIn.Provider.ExternalSearch.DnB;
+
+public class DnBSearchProviderProvider : ProviderBase, IExtendedProviderMetadata, IExternalSearchProviderProvider
 {
-    public class DnBSearchProviderProvider : ProviderBase, IExtendedProviderMetadata, IExternalSearchProviderProvider
+    public IExternalSearchProvider ExternalSearchProvider { get; }
+
+    public DnBSearchProviderProvider([System.Diagnostics.CodeAnalysis.NotNull] ApplicationContext appContext) : base(appContext, GetMetaData())
     {
-        public IExternalSearchProvider ExternalSearchProvider { get; }
+        ExternalSearchProvider = appContext.Container.ResolveAll<IExternalSearchProvider>().Single(n => n.Id == DnBExternalSearchProvider.ProviderId);
+    }
 
-        public DnBSearchProviderProvider([System.Diagnostics.CodeAnalysis.NotNull] ApplicationContext appContext) : base(appContext, GetMetaData())
+    private static IProviderMetadata GetMetaData()
+    {
+        return new ProviderMetadata
         {
-            ExternalSearchProvider = appContext.Container.ResolveAll<IExternalSearchProvider>().Single(n => n.Id == DnBExternalSearchProvider.ProviderId);
-        }
-
-        private static IProviderMetadata GetMetaData()
-        {
-            return new ProviderMetadata
-            {
-                Id = DnBExternalSearchProvider.ProviderId,
-                Name = Constants.ProviderName,
-                ComponentName = Constants.ComponentName,
-                AuthTypes = new List<string>(),
-                SupportsConfiguration = true,
-                SupportsAutomaticWebhookCreation = false,
-                SupportsWebHooks = false,
-                Type = "Enricher"
-            };
-        }
-
-        public override async Task<CrawlJobData> GetCrawlJobData(ProviderUpdateContext context, IDictionary<string, object> configuration, Guid organizationId, Guid userId, Guid providerDefinitionId)
-        {
-            if (configuration == null)
-                throw new ArgumentNullException(nameof(configuration));
-
-            var result = new DnBExternalSearchJobData(configuration);
-
-            return await Task.FromResult(result);
-        }
-
-        public override Task<bool> TestAuthentication(ProviderUpdateContext context, IDictionary<string, object> configuration, Guid organizationId, Guid userId, Guid providerDefinitionId)
-        {
-            return Task.FromResult(true);
-        }
-
-        public override Task<ExpectedStatistics> FetchUnSyncedEntityStatistics(Core.ExecutionContext context, IDictionary<string, object> configuration, Guid organizationId, Guid userId, Guid providerDefinitionId)
-        {
-            throw new NotImplementedException();
-        }
-
-
-        public override async Task<IDictionary<string, object>> GetHelperConfiguration(ProviderUpdateContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId)
-        {
-            if (jobData is DnBExternalSearchJobData result)
-            {
-                return await Task.FromResult(result.ToDictionary());
-            }
-
-            throw new InvalidOperationException($"Unexpected data type for {nameof(DnBExternalSearchJobData)}, {jobData.GetType()}");
-        }
-
-        public override Task<IDictionary<string, object>> GetHelperConfiguration(ProviderUpdateContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId, string folderId)
-        {
-            return GetHelperConfiguration(context, jobData, organizationId, userId, providerDefinitionId);
-        }
-
-        public override Task<AccountInformation> GetAccountInformation(Core.ExecutionContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId)
-        {
-            return Task.FromResult(new AccountInformation(providerDefinitionId.ToString(), providerDefinitionId.ToString()));
-        }
-
-        public override string Schedule(DateTimeOffset relativeDateTime, bool webHooksEnabled)
-        {
-            return $"{relativeDateTime.Minute} 0/23 * * *";
-        }
-
-        public override Task<IEnumerable<WebHookSignature>> CreateWebHook(Core.ExecutionContext context, CrawlJobData jobData, IWebhookDefinition webhookDefinition, IDictionary<string, object> config)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override Task<IEnumerable<WebhookDefinition>> GetWebHooks(Core.ExecutionContext context)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override Task DeleteWebHook(Core.ExecutionContext context, CrawlJobData jobData, IWebhookDefinition webhookDefinition)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override Task<CrawlLimit> GetRemainingApiAllowance(Core.ExecutionContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId)
-        {
-            if (jobData == null) throw new ArgumentNullException(nameof(jobData));
-            return Task.FromResult(new CrawlLimit(-1, TimeSpan.Zero));
-        }
-
-        public override IEnumerable<string> WebhookManagementEndpoints(IEnumerable<string> ids)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override bool ScheduleCrawlJobs => false;
-        public string Icon { get; } = "Resources.dnb.svg";
-        public string Domain { get; } = "https://www.dnb.com/";
-        public string About { get; } = "Dun & Bradstreet is global provider of business decisioning data and analytics.";
-        public AuthMethods AuthMethods { get; } = Constants.AuthMethods;
-        public IEnumerable<Control> Properties { get; } = new List<Control>();
-        public Guide Guide { get; } = null;
-        public new IntegrationType Type { get; } = IntegrationType.Enrichment;
-        public bool SupportsEnricherV2 => true;
-        public Dictionary<string, object> ExtraInfo { get; } = new()
-        {
-            { "autoMap", true },
-            { "useEnricherOriginEntityCode", true },
-            { "supportConfidenceScore", false }, // for UI
-            { "minConfidenceScore", 0 }, // for UI
-            { "maxConfidenceScore", 100 }, // for UI
-            { "origin", Constants.ProviderName.ToCamelCase() },
-            { "originField", string.Empty },
-            { "nameKeyField", Constants.KeyName.OrgNameKey },
-            { "vocabKeyPrefix", StaticDnBVocabulary.BusinessPartner.KeyPrefix },
-            { "autoSubmission", false },
-            { "dataSourceSetId", string.Empty },
+            Id = DnBExternalSearchProvider.ProviderId,
+            Name = Constants.ProviderName,
+            ComponentName = Constants.ComponentName,
+            AuthTypes = new List<string>(),
+            SupportsConfiguration = true,
+            SupportsAutomaticWebhookCreation = false,
+            SupportsWebHooks = false,
+            Type = "Enricher"
         };
     }
+
+    public override async Task<CrawlJobData> GetCrawlJobData(ProviderUpdateContext context, IDictionary<string, object> configuration, Guid organizationId, Guid userId, Guid providerDefinitionId)
+    {
+        if (configuration == null)
+            throw new ArgumentNullException(nameof(configuration));
+
+        var result = new DnBExternalSearchJobData(configuration);
+
+        return await Task.FromResult(result);
+    }
+
+    public override Task<bool> TestAuthentication(ProviderUpdateContext context, IDictionary<string, object> configuration, Guid organizationId, Guid userId, Guid providerDefinitionId)
+    {
+        return Task.FromResult(true);
+    }
+
+    public override Task<ExpectedStatistics> FetchUnSyncedEntityStatistics(ExecutionContext context, IDictionary<string, object> configuration, Guid organizationId, Guid userId, Guid providerDefinitionId)
+    {
+        throw new NotImplementedException();
+    }
+
+
+    public override async Task<IDictionary<string, object>> GetHelperConfiguration(ProviderUpdateContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId)
+    {
+        if (jobData is DnBExternalSearchJobData result)
+        {
+            return await Task.FromResult(result.ToDictionary());
+        }
+
+        throw new InvalidOperationException($"Unexpected data type for {nameof(DnBExternalSearchJobData)}, {jobData.GetType()}");
+    }
+
+    public override Task<IDictionary<string, object>> GetHelperConfiguration(ProviderUpdateContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId, string folderId)
+    {
+        return GetHelperConfiguration(context, jobData, organizationId, userId, providerDefinitionId);
+    }
+
+    public override Task<AccountInformation> GetAccountInformation(ExecutionContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId)
+    {
+        return Task.FromResult(new AccountInformation(providerDefinitionId.ToString(), providerDefinitionId.ToString()));
+    }
+
+    public override string Schedule(DateTimeOffset relativeDateTime, bool webHooksEnabled)
+    {
+        return $"{relativeDateTime.Minute} 0/23 * * *";
+    }
+
+    public override Task<IEnumerable<WebHookSignature>> CreateWebHook(ExecutionContext context, CrawlJobData jobData, IWebhookDefinition webhookDefinition, IDictionary<string, object> config)
+    {
+        throw new NotImplementedException();
+    }
+
+    public override Task<IEnumerable<WebhookDefinition>> GetWebHooks(ExecutionContext context)
+    {
+        throw new NotImplementedException();
+    }
+
+    public override Task DeleteWebHook(ExecutionContext context, CrawlJobData jobData, IWebhookDefinition webhookDefinition)
+    {
+        throw new NotImplementedException();
+    }
+
+    public override Task<CrawlLimit> GetRemainingApiAllowance(ExecutionContext context, CrawlJobData jobData, Guid organizationId, Guid userId, Guid providerDefinitionId)
+    {
+        if (jobData == null) throw new ArgumentNullException(nameof(jobData));
+        return Task.FromResult(new CrawlLimit(-1, TimeSpan.Zero));
+    }
+
+    public override IEnumerable<string> WebhookManagementEndpoints(IEnumerable<string> ids)
+    {
+        throw new NotImplementedException();
+    }
+
+    public override bool ScheduleCrawlJobs => false;
+    public string Icon => "Resources.dnb.svg";
+    public string Domain => "https://www.dnb.com/";
+    public string About => "Dun & Bradstreet is global provider of business decisioning data and analytics.";
+    public AuthMethods AuthMethods { get; } = Constants.AuthMethods;
+    public IEnumerable<Control> Properties { get; } = new List<Control>();
+    public Guide Guide => null;
+    public new IntegrationType Type => IntegrationType.Enrichment;
+    public bool SupportsEnricherV2 => true;
+    public Dictionary<string, object> ExtraInfo { get; } = new()
+    {
+        { "autoMap", true },
+        { "useEnricherOriginEntityCode", true },
+        { "supportConfidenceScore", false }, // for UI
+        { "minConfidenceScore", 0 }, // for UI
+        { "maxConfidenceScore", 100 }, // for UI
+        { "origin", Constants.ProviderName.ToCamelCase() },
+        { "originField", string.Empty },
+        { "nameKeyField", Constants.KeyName.OrgNameKey },
+        { "vocabKeyPrefix", StaticDnBVocabulary.BusinessPartner.KeyPrefix },
+        { "autoSubmission", false },
+        { "dataSourceSetId", string.Empty },
+    };
 }
