--- src/ExternalSearch.Providers.Dnb/Model/AuthResponse.cs
@@ -1,13 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace CluedIn.ExternalSearch.Providers.DnB.Model
-{
-    public class AuthResponse
-    {
-        [JsonProperty("access_token")]
-        public string AccessToken { get; set; }
-
-        [JsonProperty("expiresIn")]
-        public int ExpiresIn { get; set; }
-    }
-}
