+++ src/ExternalSearch.Providers.Dnb/Model/AuthResponse/AuthResponse.cs
@@ -1,0 +1,12 @@
+﻿using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.DnB.Model.AuthResponse;
+
+public class AuthResponse
+{
+    [JsonProperty("access_token")]
+    public string AccessToken { get; set; }
+
+    [JsonProperty("expiresIn")]
+    public int ExpiresIn { get; set; }
+}
