--- src/ExternalSearch.Providers.Dnb/Model/Result.cs
@@ -1,9 +1,0 @@
-﻿using CluedIn.ExternalSearch.Providers.DnB.Models;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace CluedIn.ExternalSearch.Providers.DnB.Model
-{
-   
-}
